requirements.txt
openai>=1.0.0
tiktoken
python-dotenv
pyaudio
numpy
torch
wave
requests
pygame
websockets
asyncio
funasr
funasr-onnx
fastapi
uvicorn
python-multipart 